+++ api/src/routes/ui.routes.ts
@@ -27,9 +27,11 @@ export class UIRoutes {
         this.router.get('/ui/assets', (req: Request, res: Response) => uiController.getAssets(req, res));
         this.router.get('/ui/substation-geojson', (req: Request, res: Response) => uiController.getSubstationGeoJSON(req, res));
         this.router.get('/ui/power-line-geojson', (req: Request, res: Response) => uiController.getPowerLineGeoJSON(req, res));
+        this.router.get('/ui/solar-potential', (req: Request, res: Response) => uiController.getSolarPotentialAtLocation(req, res));
         this.router.post('/ui/layer/:layerId', (req: Request, res: Response) => uiController.processLayerGeoJSON(req, res));
         this.router.post('/ui/location/analyse', (req: Request, res: Response) => uiController.analyseLocation(req, res));
         this.router.post('/ui/asset/analyse', (req: Request, res: Response) => uiController.analyseAsset(req, res));
+        this.router.post('/ui/asset/estimate', (req: Request, res: Response) => uiController.estimateAssetContribution(req, res));
         this.router.post('/ui/substations', (req: Request, res: Response) => uiController.getSubstations(req, res));
     }
 }